requirements.txt
tensorflow>=2.10
numpy
matplotlib
Pillow
ttkthemes

gradio==5.49.1
tensorflow==2.19.1
numpy
pillow
ow
